+++ Shinobi/Shinobi/Source/Particle.cpp
@@ -37,7 +37,7 @@ bool Particle::Update()
 		// If the particle has a specific lifetime, check when it has to be destroyed
 		if (lifetime > 0)
 		{
-			if (frameCount >= lifetime)
+			if (static_cast<unsigned int>(frameCount) >= lifetime)
 				ret = false;
 		}
 		// Otherwise the particle is destroyed when the animation is finished
